feat: improved file structure and fix prettier issues
