+++ src/components/NavBar/Nav.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { NavContainer, NavItem } from './styles';
 import { NavData } from '@/config/content/NavData';
-import { NavbarLink } from '@/components/shared/Typography/Links';
-import Link from 'next/link';
 
 export const NavBar = () => {
   return (
