+++ src/components/eventsCarousel/eventCardComponents/Parent.Style.jsx
