+++ src/components/eventsCarousel/utils/ImageContainer.style.jsx
