+++ src/components/eventsCarousel/wrapperComponents/EventWrapper.styles.jsx
